Flattened vision code
