--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Enhancement/Subsystems/Vision/DetectMarker/MarkerLocation.java
@@ -1,12 +1,0 @@
-package org.firstinspires.ftc.teamcode.Enhancement.Subsystems.Vision.DetectMarker;
-
-/**
- * The Marker Location is {@link #SEARCHING} by default.
- */
-public enum MarkerLocation {
-    LEFT,
-    MIDDLE,
-    RIGHT,
-    NOT_FOUND,
-    SEARCHING
-}
