+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Enhancement/Subsystems/Vision/DetectMarkerPipeline.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Enhancement.Subsystems.Vision.DetectMarker;
+package org.firstinspires.ftc.teamcode.Enhancement.Subsystems.Vision;
 
 import org.firstinspires.ftc.teamcode.Enhancement.Config.MainConfig;
 import org.firstinspires.ftc.teamcode.Enhancement.Robot;
@@ -20,6 +20,14 @@ import org.openftc.easyopencv.OpenCvPipeline;
  * @see Vision
  */
 public class DetectMarkerPipeline extends OpenCvPipeline {
+    public enum MarkerLocation {
+        LEFT,
+        MIDDLE,
+        RIGHT,
+        NOT_FOUND,
+        SEARCHING
+    }
+
     private final AllianceColor allianceColor = MainConfig.getAllianceColor();
     private final Rect LEFT_RECT = new Rect(
             new Point(60, 35),
