--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Tests/TestsUtil/AssertException.java
@@ -1,10 +1,0 @@
-package org.firstinspires.ftc.teamcode.Tests.TestsUtil;
-
-/**
- * Meant for {@link Test#Assert(Object, Object)}
- */
-public class AssertException extends Exception {
-    public AssertException(Object first, Object second) {
-        super(first.toString() + " is not equal to " + second.toString());
-    }
-}
