--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Tests/UnitTests/Subsystems/testRobot.java
@@ -1,11 +1,0 @@
-package org.firstinspires.ftc.teamcode.Tests.UnitTests.Subsystems;
-
-
-/**
- * Tests basic robot functions
- */
-public class testRobot {
-    public testRobot() {
-
-    }
-}
